+++ modules/nrfx/drivers/include/nrfx_comp.h
@@ -218,9 +218,9 @@ uint32_t nrfx_comp_sample(void);
  *
  * @return Address of the given COMP task.
  */
-__STATIC_INLINE uint32_t nrfx_comp_task_address_get(nrf_comp_task_t task)
+__STATIC_INLINE uintptr_t nrfx_comp_task_address_get(nrf_comp_task_t task)
 {
-    return (uint32_t)nrf_comp_task_address_get(task);
+    return (uintptr_t)nrf_comp_task_address_get(task);
 }
 
 /**
@@ -230,9 +230,9 @@ __STATIC_INLINE uint32_t nrfx_comp_task_address_get(nrf_comp_task_t task)
  *
  * @return Address of the given COMP event.
  */
-__STATIC_INLINE uint32_t nrfx_comp_event_address_get(nrf_comp_event_t event)
+__STATIC_INLINE uintptr_t nrfx_comp_event_address_get(nrf_comp_event_t event)
 {
-    return (uint32_t)nrf_comp_event_address_get(event);
+    return (uintptr_t)nrf_comp_event_address_get(event);
 }
 
 /** @} */
